- save some CPU by using only float as the DSP accuracy. We can do this   because even floats give us > 100 dB SNR -- that simply has to be enough. - implement tubeamp.c convolution with SSE where available. The floats are   not one bit faster than the integer version of the code, though, but   SSE2 should do the convolution bit faster. I'll see what I can do to   optimize it further.

+++ src/pump.h
@@ -33,7 +33,7 @@
 #include "glib12-compat.h"
 
 #ifdef FLOAT_DSP
-typedef double	DSP_SAMPLE;
+typedef float	DSP_SAMPLE;
 #else
 typedef gint32	DSP_SAMPLE;
 #endif
